requirements.txt
flask==2.0.1
yt-dlp==2023.11.16
assemblyai==0.15.0
python-dotenv==0.19.0
requests==2.31.0
gunicorn==21.2.0
flask-cors==4.0.0 